+++ src/Managers/Assistants/DuplicateAssistant.php
@@ -31,7 +31,7 @@ trait DuplicateAssistant
         }
 
         try {
-            $this->authorize('create-page');
+            $this->authorizeResourcePermission('create');
         } catch (NotAllowedManagerAction $e) {
             return false;
         }